requirements.txt
aiohttp==3.11.13
python-dotenv==1.0.1
pydantic==2.10.6
psutil==7.0.0
backoff==2.2.1
PyYAML==6.0.1
